# Improve processing performance