Always iterating over all previous subtitle lines doesn't scale well and lead to an overhead of ~2.7s for a 5h 16min long podcast (total time for getting all lines was ~12.1s).

By keeping track of visible subtitle lines in a separate array the processing time was reduced to ~40ms.

For a shorter video (23min) the processing time went from ~14ms to ~5ms.

Since practically all time is now spent in stepping through subtitles and querying their start/stop time and text, none of which can be skipped, it's not possible to achieve any further meaningful speedups.